// src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        LockOut: {
            /** Id */
            id: number;
            /** Project Id */
            project_id: number | null;
            /** Period Year */
            period_year: number;
            /** Period Month */
            period_month: number;
            /** Is Locked */
            is_locked: boolean;
            /** Locked By */
            locked_by: number | null;
            /** Locked At */
            locked_at: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}